moved add event form to its own component and added toggle on btns to open and close it
